Refactor: move test journal data
