+++ test/pages/settings/dashboards/dashboard_definition_page_test.dart
@@ -16,8 +16,8 @@ import 'package:lotti/widgets/sync/imap_config_utils.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:mocktail/mocktail.dart';
 
+import '../../../journal_test_data/test_data.dart';
 import '../../../mocks/mocks.dart';
-import '../../../test_data.dart';
 import '../../../widget_test_utils.dart';
 
 void main() {
