+++ test/pages/settings/measurables/measurable_details_page_test.dart
@@ -10,8 +10,8 @@ import 'package:lotti/themes/themes_service.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:mocktail/mocktail.dart';
 
+import '../../../journal_test_data/test_data.dart';
 import '../../../mocks/mocks.dart';
-import '../../../test_data.dart';
 import '../../../widget_test_utils.dart';
 
 void main() {
