+++ test/pages/settings/measurables/measurables_page_test.dart
@@ -5,8 +5,8 @@ import 'package:lotti/get_it.dart';
 import 'package:lotti/pages/settings/measurables/measurables_page.dart';
 import 'package:lotti/themes/themes_service.dart';
 
+import '../../../journal_test_data/test_data.dart';
 import '../../../mocks/mocks.dart';
-import '../../../test_data.dart';
 import '../../../widget_test_utils.dart';
 
 void main() {
